Перенес вспомогательные функции в utils.php
