+++ utils/utils.php
@@ -109,3 +109,32 @@ function array_order_by_key($array, $ordered_keys) {
 
     return $result;
 }
+
+// ToDo
+// Как это можно реализовать по другому?
+// Далеко не элегантное решение.
+// ключи в $form_data и $errors должны совпадать.
+// постоянно нужно передавать $form_data, $errors... По идее нужно класс сделать.
+function show_form_data($key, $form_data, $errors) 
+{
+  $result = '';
+
+  if(!isset($errors[$key]) && isset($form_data[$key])) {
+    $result = $form_data[$key];
+  }
+    
+  return $result;
+}
+
+// ToDo
+// Та же история. 
+function show_error($key, $errors) 
+{
+  $result = '';
+
+  if(isset($errors[$key])) {
+    $result = $errors[$key];
+  }
+
+  return $result;
+}
